requirement.txt
loguru
joblib
tensorboard
tensorboardX
matplotlib
thop
ruamel.yaml

ml_collections
tqdm
scikit-learn
einops
timm
opencv-python
imgaug

monai
nibabel
lifelines
SimpleITK
pydicom
dicompyler-core
